Refactored raw data reader into separate function
